The Mesoscopic Membrane Manifold Stability Theorem
This paper introduces the concept of a microscopic-mesoscopic-macroscopic
multi-scale dynamic system for the first time, addressing chaotic phenomena
such as turbulence, combustion, and axial compressor. The limitations of applying Lyapunov stability theory to such multi-scale dynamic systems are analyzed.
Based on this analysis, the concept of the mesoscopic membrane manifold and
the corresponding stability theorem are proposed. The paper posits that instability in a microscopic-mesoscopic-macroscopic multi-scale dynamic system arises
from the transmission of microscopic disorder to the macroscopic level, with the
mesoscopic force field and mesoscopic membrane manifold being decisive factors
in this process. A mesoscopic membrane stability analysis theory for a class of
microscopic-mesoscopic-macroscopic multi-scale dynamic systems is presented.
This theory addresses the limitations of Lyapunov stability theory in the context
of multi-scale dynamic systems, effectively expanding the methods available for
stability analysis of complex mesoscopic energy systems. Additionally, it provides
new perspectives for stability analysis, mechanical calculations, and the design
of controllers for active stability control in aeroengine compressors.